Move C# and VB SyntaxFacts and SyntaxKinds to a more specific namespace
